01_Ecommerce_Analytics_MultiCloud_Pipeline/pipeline/requirements.txt
pandas==2.2.2
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pyarrow==16.1.0
google-cloud-bigquery==3.25.0
